Moved prescriptor to its own class and moved trainer out of prescriptor
